// frontend/server/generated/bucket.ts
export interface BucketAliasItem {
  bucket_id: string;
  global_alias?: string | undefined;
  local_alias?: LocalAliasInput | undefined;
}

export interface BucketKeyPermissionRequest {
  items: BucketKeyPermissionItem[];
}

export interface BucketKeyPermissionItem {
  bucket_id: string;
  access_key_id: string;
  permissions: BucketKeyPermissions | undefined;
}
// ...
export interface BucketKeyPermissions {
  read: boolean;
  write: boolean;
  owner: boolean;
}

export interface Quotas {
  max_size?: number | undefined;
  max_objects?: number | undefined;
}

export interface WebsiteConfig {
  index_document?: string | undefined;
  error_document?: string | undefined;
}

export interface LocalAliasInput {
  access_key_id: string;
  alias: string;
  permissions?: BucketKeyPermissions | undefined;
}
// ...
export const BucketAliasItem: MessageFns<BucketAliasItem> = {
  encode(message: BucketAliasItem, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.bucket_id !== "") {
      writer.uint32(10).string(message.bucket_id);
    }
    if (message.global_alias !== undefined) {
      writer.uint32(18).string(message.global_alias);
    }
    if (message.local_alias !== undefined) {
      LocalAliasInput.encode(message.local_alias, writer.uint32(26).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): BucketAliasItem {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseBucketAliasItem();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.bucket_id = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.global_alias = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.local_alias = LocalAliasInput.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): BucketAliasItem {
    return {
      bucket_id: isSet(object.bucketId)
        ? globalThis.String(object.bucketId)
        : isSet(object.bucket_id)
        ? globalThis.String(object.bucket_id)
        : "",
      global_alias: isSet(object.globalAlias)
        ? globalThis.String(object.globalAlias)
        : isSet(object.global_alias)
        ? globalThis.String(object.global_alias)
        : undefined,
      local_alias: isSet(object.localAlias)
        ? LocalAliasInput.fromJSON(object.localAlias)
        : isSet(object.local_alias)
        ? LocalAliasInput.fromJSON(object.local_alias)
        : undefined,
    };
  },

  toJSON(message: BucketAliasItem): unknown {
    const obj: any = {};
    if (message.bucket_id !== "") {
      obj.bucketId = message.bucket_id;
    }
    if (message.global_alias !== undefined) {
      obj.globalAlias = message.global_alias;
    }
    if (message.local_alias !== undefined) {
      obj.localAlias = LocalAliasInput.toJSON(message.local_alias);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<BucketAliasItem>, I>>(base?: I): BucketAliasItem {
    return BucketAliasItem.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<BucketAliasItem>, I>>(object: I): BucketAliasItem {
    const message = createBaseBucketAliasItem();
    message.bucket_id = object.bucket_id ?? "";
    message.global_alias = object.global_alias ?? undefined;
    message.local_alias = (object.local_alias !== undefined && object.local_alias !== null)
      ? LocalAliasInput.fromPartial(object.local_alias)
      : undefined;
    return message;
  },
};
// ...
export const BucketKeyPermissionItem: MessageFns<BucketKeyPermissionItem> = {
  encode(message: BucketKeyPermissionItem, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.bucket_id !== "") {
      writer.uint32(10).string(message.bucket_id);
    }
    if (message.access_key_id !== "") {
      writer.uint32(18).string(message.access_key_id);
    }
    if (message.permissions !== undefined) {
      BucketKeyPermissions.encode(message.permissions, writer.uint32(26).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): BucketKeyPermissionItem {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseBucketKeyPermissionItem();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.bucket_id = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.access_key_id = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.permissions = BucketKeyPermissions.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): BucketKeyPermissionItem {
    return {
      bucket_id: isSet(object.bucketId)
        ? globalThis.String(object.bucketId)
        : isSet(object.bucket_id)
        ? globalThis.String(object.bucket_id)
        : "",
      access_key_id: isSet(object.accessKeyId)
        ? globalThis.String(object.accessKeyId)
        : isSet(object.access_key_id)
        ? globalThis.String(object.access_key_id)
        : "",
      permissions: isSet(object.permissions) ? BucketKeyPermissions.fromJSON(object.permissions) : undefined,
    };
  },

  toJSON(message: BucketKeyPermissionItem): unknown {
    const obj: any = {};
    if (message.bucket_id !== "") {
      obj.bucketId = message.bucket_id;
    }
    if (message.access_key_id !== "") {
      obj.accessKeyId = message.access_key_id;
    }
    if (message.permissions !== undefined) {
      obj.permissions = BucketKeyPermissions.toJSON(message.permissions);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<BucketKeyPermissionItem>, I>>(base?: I): BucketKeyPermissionItem {
    return BucketKeyPermissionItem.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<BucketKeyPermissionItem>, I>>(object: I): BucketKeyPermissionItem {
    const message = createBaseBucketKeyPermissionItem();
    message.bucket_id = object.bucket_id ?? "";
    message.access_key_id = object.access_key_id ?? "";
    message.permissions = (object.permissions !== undefined && object.permissions !== null)
      ? BucketKeyPermissions.fromPartial(object.permissions)
      : undefined;
    return message;
  },
};
// ...
export const LocalAliasInput: MessageFns<LocalAliasInput> = {
  encode(message: LocalAliasInput, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.access_key_id !== "") {
      writer.uint32(10).string(message.access_key_id);
    }
    if (message.alias !== "") {
      writer.uint32(18).string(message.alias);
    }
    if (message.permissions !== undefined) {
      BucketKeyPermissions.encode(message.permissions, writer.uint32(26).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): LocalAliasInput {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseLocalAliasInput();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.access_key_id = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.alias = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.permissions = BucketKeyPermissions.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): LocalAliasInput {
    return {
      access_key_id: isSet(object.accessKeyId)
        ? globalThis.String(object.accessKeyId)
        : isSet(object.access_key_id)
        ? globalThis.String(object.access_key_id)
        : "",
      alias: isSet(object.alias) ? globalThis.String(object.alias) : "",
      permissions: isSet(object.permissions) ? BucketKeyPermissions.fromJSON(object.permissions) : undefined,
    };
  },

  toJSON(message: LocalAliasInput): unknown {
    const obj: any = {};
    if (message.access_key_id !== "") {
      obj.accessKeyId = message.access_key_id;
    }
    if (message.alias !== "") {
      obj.alias = message.alias;
    }
    if (message.permissions !== undefined) {
      obj.permissions = BucketKeyPermissions.toJSON(message.permissions);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<LocalAliasInput>, I>>(base?: I): LocalAliasInput {
    return LocalAliasInput.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<LocalAliasInput>, I>>(object: I): LocalAliasInput {
    const message = createBaseLocalAliasInput();
    message.access_key_id = object.access_key_id ?? "";
    message.alias = object.alias ?? "";
    message.permissions = (object.permissions !== undefined && object.permissions !== null)
      ? BucketKeyPermissions.fromPartial(object.permissions)
      : undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}